// src/models/users/DTO/User.DTO.ts
export default interface UserDTO {
   readonly name: string;
   readonly lastname: string;
   readonly username: string;
   readonly email: string;
   readonly password: string;
   readonly phone: string;
   readonly address: string;
   readonly city: string;
   readonly country: string;
   readonly avatar: string;
}